Extract TimelineState into types.ts to avoid store<->modules cycle
- 品質レビュー指摘対応: record/prune/query が親モジュール store.ts の
  型を逆参照していた（store -> record/prune/query -> store の
  循環構造）ため、TimelineState を新設の types.ts へ切り出し
- store.ts / record.ts / prune.ts / query.ts はいずれも types.ts から
  一方向に import する形へ変更（機械的な型定義の移動のみ、ロジック変更なし）
