+++ apps/server/src/state-timeline/types.ts
@@ -1,0 +1,8 @@
+import type { TimelineEvent } from "./timeline-restore";
+
+export type TimelineState = {
+  eventsByPane: Map<string, TimelineEvent[]>;
+  sequence: number;
+  now: () => Date;
+  retentionMs: number;
+};
